requirements.txt
boto3==1.35.52
pandas==2.2.3
python-dotenv==1.0.1
snowflake-connector-python==3.12.2
SQLAlchemy==1.4.54